feat: migrate to webpack4
The following PR introduces a new structure to the gulp tasks. 

Previously react tasks was consuming typescript tasks, which were consuming the common tasks. This structure has proven to be not the most optimal one as we were unable to benefit from the @dev channel of common tasks package. 

The current setup is directly consuming its tasks from both the typescript and common tasks packages. 

The tasks used by the common tasks packages are related to the docs server, while the ones from the typescript tasks are for the API generation. 

All other tasks are hoisted to this package thus giving us more control over their execution and the versioning of the utilities used in them.

BREAKING CHANGE: Migrate to webpack4